{"type":"TI", "runs":"1"}

+++ FunWithDivision.java
@@ -6,10 +6,10 @@ import acm.program.*;
 
 public class FunWithDivision extends ConsoleProgram {
 	public void run() {
-		int x = 5;
-		int y = 3;
+		double x = 5;
+		double y = 3;
 		
-		int ratio = x / y;
+		double ratio = x / y;
 		
 		println("x / y = " + ratio);
 	}
